requirements.txt
tensorflow_gpu==1.12.1
numpy==1.16.2
scipy==1.2.1
tqdm==4.31.1
scikit_image==0.14.2
matplotlib==3.0.3
imgaug==0.2.8
opencv_python==4.0.0.21
